# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ascend API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: number;
  email: string;
}

export interface SignupInput {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface ApiError {
  error: string;
}

export interface UserProfile {
  id: number;
  name: string;
  age: number;
  gender: string;
  heightCm: number;
  currentWeightKg: number;
  goalWeightKg: number;
  bodyType: string;
  goals: string[];
  /** @nullable */
  targetDate?: string | null;
  fitnessLevel: string;
  gymAccess: string;
  /** @nullable */
  equipment?: string | null;
  workoutDaysPerWeek: number;
  /** @nullable */
  preferredWorkoutTime?: string | null;
  wakeTime: string;
  sleepTime: string;
  sleepQuality?: number;
  energyLevel?: number;
  stressLevel?: number;
  /** @nullable */
  workSchedule?: string | null;
  /** @nullable */
  averageDailySteps?: number | null;
  /** @nullable */
  allergies?: string | null;
  /** @nullable */
  dislikedFoods?: string | null;
  /** @nullable */
  dietStyle?: string | null;
  /** @nullable */
  foodBudget?: string | null;
  mealsPerDay?: number;
  waterIntakeLiters?: number;
  /** @nullable */
  caffeineUse?: string | null;
  /** @nullable */
  screenTimeBeforeBed?: string | null;
  skinConcerns?: string[];
  digestionConcerns?: string[];
  /** @nullable */
  biggestStruggle?: string | null;
  createdAt: string;
}

export interface UserProfileInput {
  name: string;
  age: number;
  gender: string;
  heightCm: number;
  currentWeightKg: number;
  goalWeightKg: number;
  bodyType: string;
  goals: string[];
  targetDate?: string;
  fitnessLevel: string;
  gymAccess: string;
  equipment?: string;
  workoutDaysPerWeek: number;
  preferredWorkoutTime?: string;
  wakeTime: string;
  sleepTime: string;
  sleepQuality?: number;
  energyLevel?: number;
  stressLevel?: number;
  workSchedule?: string;
  averageDailySteps?: number;
  allergies?: string;
  dislikedFoods?: string;
  dietStyle?: string;
  foodBudget?: string;
  mealsPerDay?: number;
  waterIntakeLiters?: number;
  caffeineUse?: string;
  screenTimeBeforeBed?: string;
  skinConcerns?: string[];
  digestionConcerns?: string[];
  biggestStruggle?: string;
}

export interface UserProfileUpdate {
  currentWeightKg?: number;
  sleepQuality?: number;
  energyLevel?: number;
  stressLevel?: number;
}

export interface Plan {
  id: number;
  userId: number;
  goalType: string;
  calorieTarget: number;
  proteinTargetG: number;
  waterTargetL: number;
  stepsTarget: number;
  sleepTargetHours: number;
  weeklyPace: string;
  workoutSchedule: string;
  keyHabits: string[];
  coachNotes: string;
  /** @nullable */
  warnings?: string | null;
  createdAt: string;
}

export interface Workout {
  id: number;
  userId: number;
  name: string;
  type: string;
  completedAt: string;
  durationMinutes: number;
  /** @nullable */
  notes?: string | null;
}

export interface WorkoutInput {
  name: string;
  type: string;
  durationMinutes: number;
  notes?: string;
}

export interface Exercise {
  name: string;
  sets: number;
  reps: string;
  restSeconds: number;
  coachTip: string;
}

export interface PlannedWorkout {
  day: string;
  name: string;
  type: string;
  exercises: Exercise[];
}

export interface Meal {
  id: number;
  userId: number;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  loggedAt: string;
  coachFeedback: string;
  score: number;
  quality: string;
  /** @nullable */
  whatWasGood?: string | null;
  /** @nullable */
  whatWasBad?: string | null;
  /** @nullable */
  whatToFixNext?: string | null;
}

/**
 * At least one of description or imageUrl must be provided.
 */
export interface MealInput {
  description?: string;
  /** Base64 data URL of the meal photo */
  imageUrl?: string;
}

export interface JournalEntry {
  id: number;
  userId: number;
  date: string;
  followedSchedule: boolean;
  hitProtein: boolean;
  stayedNearCalories: boolean;
  workedOut: boolean;
  drankWater: boolean;
  sleptOnTime: boolean;
  energyRating: number;
  skinBloatingRating: number;
  /** @nullable */
  whatWentWrong?: string | null;
  biggestWin: string;
  /** @nullable */
  needHelpWith?: string | null;
  createdAt: string;
}

export interface JournalEntryInput {
  followedSchedule: boolean;
  hitProtein: boolean;
  stayedNearCalories: boolean;
  workedOut: boolean;
  drankWater: boolean;
  sleptOnTime: boolean;
  energyRating: number;
  skinBloatingRating: number;
  whatWentWrong?: string;
  biggestWin: string;
  needHelpWith?: string;
}

export interface CoachReview {
  id: number;
  userId: number;
  date: string;
  dailyScore: number;
  biggestWin: string;
  biggestMistake: string;
  whatSlowedProgress: string;
  exactFixForTomorrow: string;
  onPace: boolean;
  strictCoachMessage: string;
  createdAt: string;
}

export interface WeighIn {
  id: number;
  userId: number;
  weightKg: number;
  weekNumber: number;
  adjustment: string;
  coachMessage: string;
  loggedAt: string;
}

export interface WeighInInput {
  weightKg: number;
  notes?: string;
}

export interface ScheduleItem {
  time: string;
  activity: string;
  type: string;
  /** @nullable */
  notes?: string | null;
}

export interface DailySchedule {
  date: string;
  items: ScheduleItem[];
  todaysMission: string;
}

export interface ChatMessageInput {
  message: string;
}

export interface ChatResponse {
  reply: string;
  timestamp: string;
}

export interface ChatMessage {
  id: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ProgressSummary {
  currentWeightKg: number;
  goalWeightKg: number;
  startWeightKg: number;
  progressPercent: number;
  dayStreak: number;
  totalWorkouts: number;
  avgDailyScore: number;
  weeklyWeighIns: WeighIn[];
  recentMeals: Meal[];
}

export interface MissionStreak {
  currentStreak: number;
  longestStreak: number;
  lastActiveDate: string;
}