# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ascend API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workouts:
    get:
      operationId: listWorkouts
      tags: [workouts]
      summary: List all workout sessions
      responses:
        "200":
          description: List of workouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"
    post:
      operationId: createWorkout
      tags: [workouts]
      summary: Log a completed workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
...
  /workouts/{id}/complete:
    patch:
      operationId: completeWorkout
      tags: [workouts]
      summary: Mark workout as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
...
components:
  schemas:
...
    Workout:
      type: object
      required:
        - id
        - userId
        - name
        - type
        - completedAt
        - durationMinutes
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        type:
          type: string
        completedAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        notes:
          type: ["string", "null"]
    WorkoutInput:
      type: object
      required:
        - name
        - type
        - durationMinutes
      properties:
        name:
          type: string
        type:
          type: string
        durationMinutes:
          type: integer
        notes:
          type: string
...